fix: product quantity manipulation on the home page
Therefore, it was necessary to create a Product component specifically for the Home page, containing
the logic for updating the quantity and adding the product to the cart.

BREAKING CHANGE: Remove the responsibility of the AddToCart component to handle the logic, remove
state handling in the product context, remove unnecessary typing in the product context.